backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.118.0",
    "fastembed>=0.7.3",
    "groq>=0.32.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.8",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "paddleocr>=3.2.0",
    "pillow>=11.3.0",
    "psycopg2>=2.9.10",
    "pymupdf>=1.26.4",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "redis[hiredis]>=6.4.0",
    "rq>=2.6.0",
    "sqlalchemy>=2.0.43",
    "sqlalchemy-utils>=0.42.0",
    "uuid>=1.30",
    "uvicorn>=0.37.0",
]
